// app/src/generated/models/Csa_contactsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_contactsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_contactsstatecode = keyof typeof Csa_contactsstatecode;
export const Csa_contactsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_contactsstatuscode = keyof typeof Csa_contactsstatuscode;

export interface Csa_contactsBase {
  csa_contactid: string;
  "csa_CustomerId@odata.bind": string;
  csa_email?: string;
  csa_name?: string;
  csa_phone?: string;
  csa_role?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_contactsstatecode;
  statuscode?: Csa_contactsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_contacts extends Csa_contactsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_customeridname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  csa_customerid?: object;
  _csa_customerid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}